% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specfile.R
\name{apply_filter_specfile}
\alias{apply_filter_specfile}
\title{Apply single filter to spectra HDF5 data file}
\usage{
apply_filter_specfile(aviris_h5, quosure)
}
\arguments{
\item{aviris_h5}{HDF5 file object}

\item{quosure}{An `rlang` `quosure` object (e.g. a `formula`)}
}
\description{
Apply single filter to spectra HDF5 data file
}
