% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ed.R
\name{exec_in_dir}
\alias{exec_in_dir}
\title{Execute a system command in a directory}
\usage{
exec_in_dir(dir, exe)
}
\arguments{
\item{dir}{Directory in which to execute command}

\item{exe}{Name of command to execute}
}
\description{
Execute a system command in a directory
}
