% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_df.R
\name{extend_df}
\alias{extend_df}
\title{Extend a one-row data.frame with more elements}
\usage{
extend_df(df, ...)
}
\arguments{
\item{df}{data.frame to extend. Must have column names and only have one row.}

\item{...}{Values with which to extend `df`. Argument names are column names in `df`.}
}
\description{
Extend a one-row data.frame with more elements
}
