% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_run.R
\name{generate_run}
\alias{generate_run}
\title{Prepare ED run inputs in a single location}
\usage{
generate_run(prefix, site_lat, site_lon, css_df, pss_df, site_df, output_dir,
  common_inputs_dir, site_met_dir, ed_exe_path, ed2in_template,
  ed2in_changes = NULL, RMDIR = FALSE)
}
\arguments{
\item{prefix}{Input file prefix, matched by ED2IN (full filename is `prefix.lat(-)DD.Dlon(-)DD.D.{css,pss,site}`}

\item{site_lat}{Site latitude (decimal degrees)}

\item{site_lon}{Site longitude (decimal degrees)}

\item{css_df}{`data.frame` contianing cohort (.css) file}

\item{pss_df}{`data.frame` containing patch (.pss) file}

\item{site_df}{`data.frame` containing site (.site) file}

\item{output_dir}{Root directory where outputs are stored}

\item{common_inputs_dir}{Path to directory containing `ed-inputs/{chd,dgd}` and `OGE2old`}

\item{site_met_dir}{Path to directory containing `ED_MET_DRIVER_HEADER` and `{month}.h5` files}

\item{ed_exe_path}{Path to ED executable, which will be linked to directory}

\item{ed2in_template}{Path to template ED2IN file, which will be copied and modified}

\item{ed2in_changes}{Optional list of additional changes to make to ED2IN, in the form `TAG = "value"`}

\item{RMDIR}{If TRUE, remove output directory and contents before starting}
}
\description{
Prepare ED run inputs in a single location
}
