% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lenna}
\alias{lenna}
\title{Photograph of Lenna}
\format{an image of class \code{cimg}}
\usage{
lenna
}
\description{
The Lenna (or Lena) picture is one of the most widely used standard test
images used for compression algorithms
}
\keyword{datasets}
