% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ed.R
\name{run_ed}
\alias{run_ed}
\title{Run ED from a given directory}
\usage{
run_ed(dir, edr_exe_name = "ed_2.1")
}
\arguments{
\item{dir}{Directory in which to execute command}

\item{edr_exe_name}{Name of ED2 executable}
}
\description{
Run ED from a given directory
}
