% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_edr.R
\name{setup_edr}
\alias{setup_edr}
\title{Setup EDR directory}
\usage{
setup_edr(dir, edr_exe_path)
}
\arguments{
\item{dir}{Directory in which to execute command}

\item{edr_exe_path}{Path to EDR executable, to be linked.}
}
\description{
Setup EDR directory
}
