% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform an image}
\usage{
transform(im, s)
}
\arguments{
\item{im}{\code{cimg} object}

\item{s}{numeric \code{1} by \code{3} vector containing the registration
parameters}
}
\value{
shifted \code{cimg} object
}
\description{
Transform an image
}
\examples{
shift(cameraman, c(1,1))
shift(cameraman, cbind(c(1,1),c(-0.5,0.5)))
}
